tests(engine-api): get payload tested, new payload on progress
