+++ packages/client/package.json
@@ -44,7 +44,7 @@
     "test": "npm run test:unit && npm run test:integration && npm run test:browser",
     "test:unit": "npm run tape -- 'test/!(integration|cli)/**/*.spec.ts' 'test/*.spec.ts'",
     "test:integration": "npm run tape -- 'test/integration/**/*.spec.ts'",
-    "test:temp": "npm run tape -- 'test/rpc/engine/forkchoiceUpdatedV1.spec.ts'",
+    "test:temp": "npm run tape -- 'test/rpc/engine/*.spec.ts'",
     "test:cli": "npm run tape -- 'test/cli/*.spec.ts'",
     "test:browser": "karma start karma.conf.js",
     "tsc": "../../config/cli/ts-compile.sh"
