+++ packages/client/test/rpc/engine/getPayloadV1.spec.ts
@@ -1,7 +1,9 @@
 import tape from 'tape'
 import { INVALID_PARAMS } from '../../../lib/rpc/error-code'
-import { params, baseRequest, baseSetup } from '../helpers'
+import { params, baseRequest, baseSetup, setupChain } from '../helpers'
 import { checkError } from '../util'
+import genesisJSON from '../../testdata/post-merge.json'
+import { validPayload } from './forkchoiceUpdatedV1.spec'
 
 const method = 'engine_getPayloadV1'
 
@@ -23,4 +25,22 @@ tape(`${method}: call with unknown payloadId`, async (t) => {
   const req = params(method, ['0x123'])
   const expectRes = checkError(t, -32001, 'Unknown payload')
   await baseRequest(t, server, req, 200, expectRes)
+})
+
+tape(`${method}: call with known payload`, async (t) => {
+  const { server } = await setupChain(genesisJSON, 'post-merge', { engine: true })
+  const forkchoiceUpdateRequest = params('engine_forkchoiceUpdatedV1', validPayload)
+  let payloadId
+  const expectRes = (res: any) => {
+    payloadId = res.body.result.payloadId
+  }
+  await baseRequest(t, server, forkchoiceUpdateRequest, 200, expectRes, false)
+
+  const getPayloadRequest = params(method, [payloadId])
+
+  const expectGetPayloadResponse = (res: any) => {
+    t.equal(res.body.result.blockNumber, '0x1')
+  }
+
+  await baseRequest(t, server, getPayloadRequest, 200, expectGetPayloadResponse)
 })
